requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
mysql-connector-python==9.1.0
mysqlclient==2.2.6
pillow==11.0.0
PyMySQL==1.1.1
sqlparse==0.5.2
tzdata==2024.2
gunicorn==20.1.0
